resources/kernel-visualization/requirements.txt
pygame==2.5.2
Pillow==10.1.0
numpy==1.26.2